Passing tsgprm structure into gc316 driver to generated better raw image files

+++ qnx/root/operational_radar_code/ros/gc316_tcp_driver/gc314GetBufferAddress.c
@@ -44,7 +44,7 @@ unsigned long gc314GetBufferAddress(unsigned int *physical_addr,int filedes, int
 	}
 	return address;
 #else
-        return NULL;
+        return (unsigned long)NULL;
 #endif
 
 }
